--- src/components/SidebarLink.vue
@@ -1,14 +1,0 @@
-<template>
-  <div>
-    asd
-  </div>
-</template>
-
-<script>
-export default {
-  name: "sidebar-link",
-};
-</script>
-<style>
-</style>
-
