+++ src/data/words.json
@@ -12,7 +12,6 @@
       },
       "forgets_count": "0",
       "forgets_dates": []
-
     },
     {
       "en": "serf",
